Added DisposeListMemoryOnly function; fixed a bug in testSimpleList.c (it was scanning a long number and putting it into an int).

+++ SimpleList/testSimpleList.c
@@ -14,7 +14,7 @@ int main()
 {
     SimpleList aList;
     long input;
-    int *element, i;
+    long *element, i;
     InitList(&aList, 10);
     while (1) {
         printf("Enter new element to add, 0 to exit, -index to delete element at index:\n");
@@ -29,13 +29,13 @@ int main()
         }
         else
         {
-            element = (int*)malloc(sizeof(int));
+            element = (long*)malloc(sizeof(long));
             *element = input;
             AddElement(&aList, element);
         }
         printf("Current list elements are:\n");
         for (i = 0; i < aList.currentNum; ++i) {
-            printf("%d ", *((int*)GetElementAt(&aList, i)));
+            printf("%ld ", *((long*)GetElementAt(&aList, i)));
         }
         printf("\n");
     }
